requirements.txt
appdirs==1.4.4
asgiref==3.4.1
backports.zoneinfo==0.2.1
black==21.7b0
boto3==1.20.24
botocore==1.23.24
certifi==2020.12.5
click==8.0.1
cloudinary==1.25.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0
django-cors-headers==3.7.0
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.1.0
jmespath==0.10.0
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.8.6
PyJWT==2.0.1
python-dateutil==2.8.2
python-decouple==3.5
python-http-client==3.3.4
pytz==2021.1
regex==2021.7.6
s3transfer==0.5.0
sendgrid==6.9.2
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==2.0.3
static3==0.7.0
tomli==1.2.0
typing_extensions==4.0.0
urllib3==1.26.4
whitenoise==5.2.0
